Revert "parser.sp теперь size_t"
This reverts commit 5c5d8465ed878137566632af8adb6e019de6428e.

+++ libs/compiler/parser.h
@@ -52,7 +52,7 @@ typedef struct parser
 	int stackop[100];
 	int stackoperands[100];
 	int stacklog[100];
-	size_t sp;
+	int sp;
 	int sopnd;
 	int anst;
 	int ansttype;
